fixed timestamp type to prevent overflows

+++ CryptoApp.Infrastructure/API/CoinCap/Responses/CoinCapGetAssetsResponse.cs
@@ -2,5 +2,5 @@
 
 public record  CoinCapGetAssetsResponse(
     CoinCapAsset[] Data,
-    int Timestamp
+    long Timestamp
 );
